refactor(src/game): review changes
* move enum ScreenType
* delete typings/index.ts
