+++ packages/client/src/pages/UserProfile/UserProfile.tsx
@@ -8,8 +8,8 @@ import { ButtonVariant } from '../../components/Button/data';
 import { Form } from '../../components/Form';
 import { FieldList } from '../../components/Form/FieldList';
 import { API_ENDPOINTS, DEFAULT_AVATAR } from '../../config/constants';
-import { buildPath, determineAPIHost } from '../../utils/HTTP';
 import { authSelectors, profileSelectors, profileThunks, useAppDispatch, useAppSelector } from '../../store';
+import { buildPath, determineAPIHost } from '../../utils/HTTP';
 import { generateMetaTags } from '../../utils/seoUtils';
 import { useValidation } from '../../utils/validation';
 import { userProfileFieldList } from './data';
