+++ packages/server/index.ts
@@ -10,11 +10,10 @@ dotenv.config();
 
 import express from 'express';
 import * as fs from 'fs';
+import { createProxyMiddleware } from 'http-proxy-middleware';
 import * as path from 'path';
 
 import { HtmlWritable } from './utils/HtmlWritable';
-
-import { createProxyMiddleware } from 'http-proxy-middleware';
 
 createClientAndConnect();
